lms-be/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9

# Configuration & Environment
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Date & Time Utilities
python-dateutil==2.8.2

# Type Hints & Validation
typing-extensions==4.9.0
email-validator==2.1.0.post1

# Production Server
gunicorn==21.2.0

# Database Drivers (Optional - for PostgreSQL)
# psycopg2-binary==2.9.9
